+++ fivetran/tests/mock/resource_connector_test.go
@@ -372,6 +372,7 @@ const (
 			"soap_uri":             "soap_uri",
 			"user_id":              "user_id",
 			"share_url":            "share_url",
+			"organization":         "organization",
 
 			"report_suites":            ["report_suite"],
 			"elements":                 ["element"],
@@ -660,6 +661,7 @@ func setupMockClientConnectorResourceConfigMapping(t *testing.T) {
 			assertKeyExistsAndHasValue(t, config, "soap_uri", "soap_uri")
 			assertKeyExistsAndHasValue(t, config, "user_id", "user_id")
 			assertKeyExistsAndHasValue(t, config, "share_url", "share_url")
+			assertKeyExistsAndHasValue(t, config, "organization", "organization")
 
 			assertKeyExists(t, config, "report_suites")
 			assertArrayItems(t, config["report_suites"].([]interface{}), append(make([]interface{}, 0), "report_suite"))
@@ -1138,6 +1140,7 @@ func TestResourceConnectorConfigMappingMock(t *testing.T) {
 				soap_uri = "soap_uri"
 				user_id = "user_id"
 				share_url = "share_url"
+				organization = "organization"
 
 				report_suites = ["report_suite"]
 				elements = ["element"]